Rebase. spliting rest tests into smaller modules
